Style: 사이드바 -> 상단바 위치 변경
