+++ client/src/components/layout/topnav/TopNav.jsx
@@ -1,33 +1,51 @@
 import React from "react";
 import "./topnav.css";
+import { useLocation, Link } from "react-router-dom";
+import { Box } from '@chakra-ui/react'
 
-import { useSelector, useDispatch } from "react-redux";
-import {connect} from "../../../redux/blockchain/blockchainActions.js";
+import topbar_items from "../../../assets/JsonData/topbar_router.json";
+import logo from "../../../logo.svg";
 
 const TopNav = () => {
 
-  const dispatch = useDispatch();
-  const blockchain = useSelector(state=>state.blockchain);
-
+  const location = useLocation();
   
-  const walletConnect = (e)=>{
-    e.preventDefault();
-    dispatch(connect());
-  }
+  const activeItem = topbar_items.findIndex(
+    (item) => item.route === location.pathname
+  );
+
   
   return (
     <div className="topnav">
-      <div className="topnav__account">
-        {/* <input value={account} /> */}
-        {blockchain.account ? blockchain.account: <button onClick={walletConnect}>지갑 연결</button>}
+      <div className="topnav__logo">
+        <img src={logo}  alt="company logo" />
       </div>
-      <div className="topnav__right">
-        <div className="topnav__right-item">icon #1</div>
-        <div className="topnav__right-item">icon #2</div>
-        <div className="topnav__right-item">icon #3</div>
+      <Box className="topnav__right">
+        {topbar_items.map((item, index) => (
+        <Link to={item.route} key={index}>
+          <TopItem
+            title={item.display_name}
+            active={index === activeItem}
+          />
+        </Link>
+      ))}
+      </Box>
+    </div>
+  );
+};
+
+const TopItem = (props) => {
+  const active = props.active ? "active" : "";
+
+  return (
+    <div className="topnav__item">
+      <div className={`topnav__item-inner ${active}`}>
+        {/* <i className={props.icon}></i> */}
+        <span>{props.title}</span>
       </div>
     </div>
   );
 };
+
 
 export default TopNav;
