+++ client/src/components/layout/topnav/topnav.css
@@ -30,3 +30,50 @@
 .topnav__right-item:hover {
   color: var(--main-color);
 }
+
+.topnav__logo > img {
+  height: 50px;
+}
+
+.topnav__logo {
+  height: 150px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.topnav__item-inner > span {
+  text-transform: capitalize;
+}
+
+.topnav__item-inner:hover {
+  color: var(--main-color);
+}
+
+.topnav__item-inner.active {
+  border-radius: var(--border-radius);
+  background-image: linear-gradient(
+    to right,
+    var(--main-color),
+    var(--second-color)
+  );
+  color: var(--txt-white);
+}
+
+
+.topnav__item {
+  padding: 0 20px;
+}
+
+.topnav__item-inner {
+  padding: 10px 15px;
+  display: flex;
+  align-items: center;
+  font-weight: 600;
+  transition: color 0.3s ease 0s;
+}
+
+.topnav__item-inner > i {
+  margin-right: 10px;
+  font-size: 1.5rem;
+}
